Elastic torchrun: new round if worker leaves/misses heartbeat
Fixes: #111646

# Problem
When a worker misses its heartbeat it is just removed from the list of participants, but in this case the rendezvous should restart as the world size changes. The agent is [responsible for this](https://github.com/georgkaleido/pytorch/blob/1eb6c4314b05ffa6888fea5b9a88592a2328c9f3/torch/distributed/elastic/agent/server/api.py#L909) as the rendezvous backend cannot trigger a new rendezvous on its own.

# Solution
In case a participant is removed, one remaining participant is added to the wait_list in order to trigger a new rendezvous on all workers

# Alternative
* Extend the rdvz backends by a field to track and expose dead_nodes and check this in the agent, would also need to extend the [`_RendezvousState`](https://github.com/georgkaleido/pytorch/blob/1eb6c4314b05ffa6888fea5b9a88592a2328c9f3/torch/distributed/elastic/rendezvous/dynamic_rendezvous.py#L259) and [change](https://github.com/georgkaleido/pytorch/blob/1eb6c4314b05ffa6888fea5b9a88592a2328c9f3/torch/distributed/elastic/rendezvous/dynamic_rendezvous.py#L443) how missed heart-beats are handled